zsh: Customize dircolors for clarity/contrast
The stock colors are just too noisy, and some of the combos are just
completely illegible (blue on green? really? not nearly enough
contrast).

I removed all extension-based colors.

I removed colors for "special" or "noteworthy" permissions/bits. Instead
I opted for underlines for "hey you should look at the permissions for
this file; it's got something interesting" (like setuid and sticky
bits).

I kept a simple color scheme (which is very similar to the stock
colors):
- Blue for directories
- White for regular files
- Yellow for device files
- Magenta for sockets
- Cyan for symlinks
- Red for orphaned links
- Green for executable files

That's it. No backgrounds. So much easier to read.