zsh: Fix cd-up widget not updating from git dirs
When Spaceship truncates directories to their git repos, using the cd-up
widget fails to update the displayed dir in the prompt. This, I believe,
is due to the prompt still thinking it's in a git repo when it
re-renders the directory. By forcibly refreshing each section in the
correct order, this no longer happens.